tor: updated to 0.2.9.9

+++ release/src/router/tor/src/or/or.h
@@ -2964,11 +2964,11 @@ typedef struct circuit_t {
   /** For what reason (See END_CIRC_REASON...) is this circuit being closed?
    * This field is set in circuit_mark_for_close and used later in
    * circuit_about_to_free. */
-  uint16_t marked_for_close_reason;
+  int marked_for_close_reason;
   /** As marked_for_close_reason, but reflects the underlying reason for
    * closing this circuit.
    */
-  uint16_t marked_for_close_orig_reason;
+  int marked_for_close_orig_reason;
 
   /** Unique ID for measuring tunneled network status requests. */
   uint64_t dirreq_id;
